chore(ci): update workflow and docker configurations to v1.6.6
This commit updates the application release workflow and related Docker configurations to use version 1.6.6 of the vegito-app/local repository. This includes changes in GitHub Action workflows for version, build, tests, and application release, as well as modifications in the Docker bake files to dynamically use the updated LOCAL_VERSION for the Robot Framework image. These updates ensure consistency across CI/CD pipelines with the latest patches and features available in version 1.6.6.